# server/docs/swagger.yaml
basePath: /api
definitions:
...
  auth.DeleteFCMTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  auth.LoginInternalRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginInternalResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
    type: object
  auth.LoginOwnerRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginOwnerResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.MeResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.AccountRole'
      username:
        type: string
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
    type: object
  auth.RegisterOwnerRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  auth.RegisterOwnerResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
    type: object
  auth.SaveFCMTokenRequest:
    properties:
      device_type:
        type: string
      token:
        type: string
    required:
    - device_type
    - token
    type: object
  core.ActionResponse:
    properties:
      success:
        type: boolean
    type: object
  models.AccountRole:
    enum:
    - manager
    - receptionist
    - doctor
    - owner
    type: string
    x-enum-varnames:
    - RoleManager
    - RoleReceptionist
    - RoleDoctor
    - RoleOwner
  models.AppointmentState:
    enum:
    - Menunggu Konfirmasi
    - Diterima
    - Ditolak
    - Menunggu Dokter
    - Dalam Penanganan
    - Selesai
    - Selesai Administrasi
    type: string
    x-enum-varnames:
    - StateWaitingConfirmation
    - StateAccepted
    - StateRejected
    - StateWaitingDoctor
    - StateInTreatment
    - StateFinished
    - StateAdminFinished
  models.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
  models.ServiceType:
    enum:
    - vaccine
    - checkup
    - treatment
    type: string
    x-enum-varnames:
    - ServiceTypeVaccine
    - ServiceTypeCheckup
    - ServiceTypeTreatment
  pet.CreatePetRequest:
    properties:
      avatar_upload_id:
        type: string
      birth_date:
        type: string
      breed:
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/models.Gender'
        enum:
        - male
        - female
      initial_medical_history:
        type: string
      name:
        type: string
      species:
        type: string
    required:
    - birth_date
    - breed
    - gender
    - name
    - species
    type: object
  pet.GetPresignedURLRequest:
    properties:
      content_type:
        enum:
        - image/jpeg
        - image/png
        - image/webp
        type: string
      file_size:
        description: Max 5MB
        maximum: 5242880
        minimum: 1
        type: integer
    required:
    - content_type
    - file_size
    type: object
  pet.GetPresignedURLResponse:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      method:
        type: string
      upload_id:
        type: string
      url:
        type: string
    type: object
  pet.MyPetResponse:
    properties:
      avatar_url:
        type: string
      birth_date:
        type: string
      id:
        type: integer
      name:
        type: string
      species:
        type: string
    type: object
  pet.PetAppointmentResponse:
    properties:
      appointment_date:
        type: string
      id:
        type: string
      service_type:
        $ref: '#/definitions/models.ServiceType'
      status:
        $ref: '#/definitions/models.AppointmentState'
    type: object
  pet.PetReminderResponse:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      service_type:
        $ref: '#/definitions/models.ServiceType'
    type: object
  pet.PetResponse:
    properties:
      avatar_url:
        type: string
      birth_date:
        type: string
      breed:
        type: string
      created_at:
        type: string
      gender:
        $ref: '#/definitions/models.Gender'
      id:
        type: integer
      initial_medical_history:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      species:
        type: string
    type: object
...
host: localhost:1323
info:
  contact: {}
  description: API for VetConnect Clinic Management System.
  title: VetConnect API
  version: "1.0"
paths:
...
  /fcm/token:
    delete:
      consumes:
      - application/json
      description: Remove the Firebase Cloud Messaging token for the current user
        session.
      parameters:
      - description: FCM Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DeleteFCMTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Delete FCM Token
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: Save or update the Firebase Cloud Messaging token for the current
        user session.
      parameters:
      - description: FCM Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SaveFCMTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      security:
      - BearerAuth: []
      summary: Save FCM Token
      tags:
      - Auth
...
  /internal/login:
    post:
      consumes:
      - application/json
      description: Login as a staff member (Manager, Receptionist, Doctor) using username
        and password.
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInternalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginInternalResponse'
      summary: Login Internal
      tags:
      - Auth
  /logout:
    post:
      consumes:
      - application/json
      description: Invalidate the refresh token (from body or cookie) and clear the
        session.
      parameters:
      - description: Logout Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ActionResponse'
      summary: Logout
      tags:
      - Auth
  /me:
    get:
      description: Get the profile information of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.MeResponse'
      security:
      - BearerAuth: []
      summary: Get Current Profile
      tags:
      - Auth
  /owner/login:
    post:
      consumes:
      - application/json
      description: Login as a pet owner using email and password.
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginOwnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginOwnerResponse'
      summary: Login Owner
      tags:
      - Auth
  /owner/register:
    post:
      consumes:
      - application/json
      description: Register a new account for a pet owner.
      parameters:
      - description: Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterOwnerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterOwnerResponse'
      summary: Register Owner
      tags:
      - Auth
  /pets:
    get:
      description: Get a list of all pets owned by the currently authenticated pet
        owner.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pet.MyPetResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get My Pets
      tags:
      - Pet
    post:
      consumes:
      - application/json
      description: Register a new pet under the currently authenticated pet owner's
        account.
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pet.CreatePetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pet.PetResponse'
      security:
      - BearerAuth: []
      summary: Create Pet
      tags:
      - Pet
  /pets/{id}:
    get:
      description: Get detailed information about a specific pet owned by the current
        user.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.PetResponse'
      security:
      - BearerAuth: []
      summary: Get Pet Detail
      tags:
      - Pet
    put:
      consumes:
      - application/json
      description: Update the profile information of an existing pet owned by the
        current user.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pet.CreatePetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.PetResponse'
      security:
      - BearerAuth: []
      summary: Update Pet
      tags:
      - Pet
  /pets/{id}/appointments:
    get:
      description: Get a list of all appointments for a specific pet owned by the
        current user.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pet.PetAppointmentResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get Pet Appointments
      tags:
      - Pet
  /pets/{id}/reminders:
    get:
      description: Get a list of all reminders for a specific pet owned by the current
        user.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pet.PetReminderResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get Pet Reminders
      tags:
      - Pet
  /pets/avatar/presigned-url:
    post:
      consumes:
      - application/json
      description: Generate a temporary presigned URL for secure avatar image upload.
      parameters:
      - description: Presigned URL Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pet.GetPresignedURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.GetPresignedURLResponse'
      security:
      - BearerAuth: []
      summary: Get Presigned URL
      tags:
      - Pet
...
  /token/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token (from body or
        cookie).
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RefreshTokenResponse'
      summary: Refresh Token
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"